柵上機能の追加記述３
削除失敗時のエラーハンドリングの修正
出品者以外がdestroyできないように記述

+++ app/controllers/items_controller.rb
@@ -42,11 +42,14 @@ def update
 end
 
 def destroy
+  if current_user.id != @item.user_id
+    redirect_to root_path
+  end
   
   if @item.destroy
     redirect_to root_path
   else 
-    render item_path(@item)
+    render :show
   end
 
 end
